adds GoRouter and two new dummy screens to navigate between
