+++ pubspec.yaml
@@ -36,6 +36,7 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   slang_flutter: ^4.8.0
+  go_router: ^16.2.4
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
